Support scopes in relationships
Rails has relaced the `:conditions` relationships' option with optional lambda/proc `scopes` for a long time already. It just makes sense for Formtastic to support them the same way it does for the `:conditions` option.

Given this code:

```ruby
class Post < ApplicationRecord
  belongs_to :author, -> { active.order(name: :asc) }
end

class Author < ApplicationRecord
  scope :active, -> { where(active: true) }
end
```

```slim
- semantic_form_for(Post.new) do |f|
  = f.input :author
```

Without this change the autogenerated select input with the list of authors will include ALL authors because it effectivelly loads the list from `Post.all`. After this change, the list will be filled with `Post.active.order(name: :asc)` because the `belongs_to` relationship includes a scope.

This seems to bring parity to the old way of applying conditions to relationships via the `:conditions` option which got deprecated back in Rails 4.